python/requirements.txt
google-cloud-storage
scipy
face_recognition
numpy
